Move GroupFactory into h.resources
Put it where all the other factories and resources are.

(GroupFactory doesn't actually return GroupResource objects yet, but it
should do.)
