+++ tests/h/routes_test.py
@@ -92,7 +92,7 @@ def test_includeme():
              traverse='/{id}'),
         call('api.profile', '/api/profile'),
         call('api.debug_token', '/api/debug-token'),
-        call('api.group_member', '/api/groups/{pubid}/members/{user}', factory='h.models.group:GroupFactory', traverse='/{pubid}'),
+        call('api.group_member', '/api/groups/{pubid}/members/{user}', factory='h.resources.GroupFactory', traverse='/{pubid}'),
         call('api.search', '/api/search'),
         call('api.users', '/api/users'),
         call('api.user', '/api/users/{username}'),
@@ -108,9 +108,9 @@ def test_includeme():
              factory='h.resources.OrganizationLogoFactory', traverse='/{pubid}'),
         call('api.groups', '/api/groups'),
         call('group_create', '/groups/new'),
-        call('group_edit', '/groups/{pubid}/edit', factory='h.models.group:GroupFactory', traverse='/{pubid}'),
-        call('group_read', '/groups/{pubid}/{slug:[^/]*}', factory='h.models.group:GroupFactory', traverse='/{pubid}'),
-        call('group_read_noslug', '/groups/{pubid}', factory='h.models.group:GroupFactory', traverse='/{pubid}'),
+        call('group_edit', '/groups/{pubid}/edit', factory='h.resources.GroupFactory', traverse='/{pubid}'),
+        call('group_read', '/groups/{pubid}/{slug:[^/]*}', factory='h.resources.GroupFactory', traverse='/{pubid}'),
+        call('group_read_noslug', '/groups/{pubid}', factory='h.resources.GroupFactory', traverse='/{pubid}'),
         call('help', '/docs/help'),
         call('onboarding', '/welcome/'),
         call('custom_onboarding', '/welcome/{slug}'),
